% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{connectedComponents}
\alias{connectedComponents}
\title{Compute Connected Components in a Graph.}
\usage{
connectedComponents(neighbours)
}
\arguments{
\item{neighbours}{An adjacency list in the form of a vector of vectors.}
}
\value{
An integer vector \code{C} such that \code{C[[i]]} is a node in
the same component as \code{i}, and \code{C[[i]]=C[[j]]} iff \code{i} and
\code{j} are in the same component.
}
\description{
This is an inner function, called by other R functions in this
package, but it is left available for the expert user or developer.
}
\seealso{
Other inner functions: 
\code{\link{encodeSolutionTree}()},
\code{\link{exploreTree}()},
\code{\link{fullyGrowTree}()},
\code{\link{growTree}()},
\code{\link{rawDominationGraphFineTuned}()},
\code{\link{resetSolutionTree}()},
\code{\link{scc}()},
\code{\link{setBound}()},
\code{\link{setThreshold}()},
\code{\link{takeGraph}()},
\code{\link{takeGroups}()},
\code{\link{takeStepScore}()}
}
\concept{inner functions}
