% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumors.R
\name{do.musette}
\alias{do.musette}
\title{Run the Whole Musette Pipeline in One Function}
\usage{
do.musette(
  matrices,
  reds,
  blues = !reds,
  blind_domination_step = c(),
  color_domination_step = c(),
  blind_distance = 5e+05,
  blind_percent = 80,
  color_distance = 5e+05,
  red_percent = 80,
  blue_percent = 80,
  chromosome = NULL,
  position = NULL,
  longname = NULL,
  pathways = NULL,
  bound = 2000,
  stepmode = "strict",
  threshold = 1
)
}
\arguments{
\item{matrices}{The alteration matrices, which columns correspond to the n samples and rows are named by the corresponding alterations}

\item{reds}{A named logical vector whose names are the samples' names. The
values are \code{TRUE} for red samples, i.e. those which our solutions
should affect.}

\item{blues}{A similar vector specifying the blue samples, those that our
solutions should not affect. It is forbidden for a sample to be red and blue
at the same time !}

\item{blind_domination_step}{A character array containing the alteration types for which blind domination must be used}

\item{color_domination_step}{A character array containing the alteration types for which color domination must be used}

\item{blind_distance}{The maximum distance accepted between two alterations' loci for a blind
domination to occur. If omitted, this parameter has no effect.}

\item{blind_percent}{The required percentage of coverage for the blind domination step (see \link{domination.graph})}

\item{color_distance}{The maximum distance accepted between two alterations' loci for a blind
domination to occur. If omitted, this parameter has no effect.}

\item{red_percent}{The required percentage of red samples for the domination step (see \link{domination.graph})}

\item{blue_percent}{The required percentage of blue samples for the domination step (see \link{domination.graph})}

\item{chromosome}{(for genomic alterations) An integer vector with the same length as \code{graph}, containing
the chromosome of each alteration. If omitted, this parameter has no effect.}

\item{position}{(for genomic alterations) An integer vector with the same length as \code{graph}, containing
the 'locus' of each alteration (e.g. in base pairs).  If omitted, this parameter has no effect.}

\item{longname}{Character vector containing the long name or description of each gene}

\item{pathways}{Array containing the pathways each gene belongs to.  NOT USED (maybe in the future ?)}

\item{bound}{The program will stop generating new solutions when this number
of solutions is reached. Note : \dQuote{equally good}
solutions in terms of the step-score are added at the same time, so that the final number of solutions possibly exceeds this
\code{bound}.}

\item{stepmode}{A string to set how the step-score is calculated. Valid values
are \code{"original"}, \code{"unique"}, \code{"bestFirst"}, \code{"strict"}.}

\item{threshold}{The maximum step-score for creating new solutions (see
details about the step-score in \link{musette-algorithm}).}
}
\value{
A list of two objects. The \code{alterations} dataframe contains the list of considered alterations.
 Each row corresponds to an alteration and columns include, for each alteration:
\itemize{
\item its \code{name}, \code{type} and the concerned \code{gene}
\item its location coded by its  \code{chromosome} and \code{position} in terms of base-pairs
\item \code{longname}: its extended name
\item the \code{pathways} it belongs to
\item \code{neighbours} the samples it is present in
\item \code{redneighbours, blueneighbours, nredneighbours, nblueneighbours}: the set of red (resp. blue) samples it is present in, and the cardinal of those sets
\item the hypergeometric score \code{hyper} and the musette score \code{score} corresponding to an alteration set reduced to this single alteration
\item \code{blind_dominators}, \code{blind_dominated}, \code{color_dominators}, \code{color_dominated}: sets of alteration that dominate or are dominated by this alteration in the blind or colored way
\item \code{blind_leader}, \code{color_leader}: alteration chosen to lead the group with dominations (it may be not be a direct dominator of the considered alteration but dominate a dominator) 
\item \code{blind_repr} (\code{color_repr}): a representative chosen in this alteration's strongly connected component of the blind domination (color domination) directed graph. All the alterations in a given strongly connected component will have the same representative
\item \code{followers}: the set of all alterations whose color_leader is this particular alteration
\item \code{merged_in}: pseudo-alteration in which a real alteration is potentially merged 
\item \code{merged_from}: real alterations merged to form a pseudo-alteration
\item \code{stepscore}: stepscore of this alteration seen as an evolution of the empty set.
}
 The \code{solutions} data.frame contains a row for each generated solution, in order of decreasing score.
Columns include :
\itemize{
\item alterations of each solution (a character vector)
\item size : the number of alterations
\item number of red samples hits
\item number of blue samples hits
\item sensitivity
\item specificity
\item total number of red samples
\item total number of blue samples
\item step-score of adding the last element of this solution
\item threshold : maximum step-score of all the steps in constructing this 
solution
\item leaf : TRUE indicates that this solution has not been extended into a larger solution; FALSE indicates that the dataframe contains at least an extended solution.
\item children-threshold : minimum step-score of extending this solution into
 a new solution
\item parent : row number of the solution that was extended into this one
\item score : score of this solution.
\item all_genes : list of all the alterations belonging to this solution or
 dominated, directly or indirectly by such an alteration
\item all_loci : list of all chromosomes of the genes in this solution
}
}
\description{
For a set of alteration matrices (for example amplifications, deletions and mutations) on the same samples partitioned into reds and blues, 
\itemize{
 \item the blind-domination and color-domination relatioships are computed using the function \link{domination.graph} 
 \item for each relationship type (blind or color), «leader alterations» are chosen, which dominate a nuber of «follower alterations»
 This information will be returned in the \code{alterations} dataframe. The procedure and goes on with the leader alterations only.
 \item the solution tree is constructed using the function \link{solutions}. Solutions are returned in the \code{solutions} dataframe.
}
}
\examples{
#Real bladder data, restricted on the 20 first solution sets including only mutations
data("tcga_bladder",package="musette")   
reds= (groups == 'basal')
names(reds)=names(groups)

ll=do.musette(matrices=matrices, reds=reds, blind_domination_step = c("ampli","dele"), 
color_domination_step = c("ampli","dele"), blind_distance=5000000, 
color_distance=5000000, blind_percent=90, red_percent=80, 
blue_percent=80, chromosome = chromosome, longname=longname,
pathways=pathways,position=position, bound=20)

View(ll$solutions)
View(ll$alterations)

}
