% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domination.R
\name{domination.graph}
\alias{domination.graph}
\title{Compute the Domination Relationship in a Graph, its Strongly Connected Components and Normal Forms}
\usage{
domination.graph(
  graph,
  reds,
  blues = !reds,
  chromosome = rep(1, length(graph)),
  locus = rep(1, length(graph)),
  distance = 5e+05,
  redPercent = 50,
  bluePercent = 50
)
}
\arguments{
\item{graph}{A named list of character vectors. Each of these vectors 
should have the name of one alteration, and contain the vector of the 
names of the samples affected by this alteration. Such a list can be 
obtained from a logical matrix using \code{\link{matrix2graph}}}

\item{reds}{A named logical vector whose names are the samples' names. The
values are \code{TRUE} for red samples, i.e. those which our solutions
should affect.}

\item{blues}{A similar vector specifying the blue samples, those that our
solutions should not affect. It is forbidden for a sample to be red and blue
at the same time !}

\item{chromosome}{(for genomic alterations) An integer vector with the same length as \code{graph}, containing
the chromosome of each alteration. If omitted, this parameter has no effect.}

\item{locus}{(for genomic alterations) An integer vector with the same length as \code{graph}, containing
the position of each alteration (e.g. in base pairs).  If omitted, this parameter has no effect.}

\item{distance}{The maximum distance accepted between two alterations' loci for a
domination to occur. If omitted, this parameter has no effect.}

\item{redPercent}{The required percentage of red samples}

\item{bluePercent}{The required percentage of blue samples}
}
\value{
A list \code{l} of 6 named lists of character vectors. If \code{A} is the name of an alteration, then :
\itemize{
\item \code{l$loose_dominators[[A]]} contains the names of alterations "loosely" dominating \code{A}.
\item \code{l$loose_dominated[[A]]} contains the names of alterations "loosely" dominated by \code{A}.
\item \code{l$strict_dominators[[A]]} contains the names of alterations "strictly" dominating \code{A}.
\item \code{l$strict_dominated[[A]]} contains the names of alterations "strictly" dominated by \code{A}.
\item \code{l$strongroot[[A]]} contains the name of a "reference" alteration in \code{A}'s strongly 
connected component (the set of alterations directly or indirectly dominated by \code{A} which also 
directly or indirectly dominate \code{A}). The list \code{l$strongroot} can be used to check whether
two alterations are in the same strongly connected component.
\item \code{l$strongnormal[[A]]} contains the name of a "normal form", a gene which dominates
\code{A} and is not dominated strictly by any other alterations.
}
}
\description{
(See details below about what "domination" means) 
Compute for every alteration A :
  \itemize{
    \item the list of alterations dominated by A "loosely" (i.e. including those dominating A reciprocally)
    \item the list of alterations dominating A "loosely" (i.e. including those dominated by A reciprocally)
    \item the list of alterations dominated by A "strictly" (i.e. excluding those dominating A)
    \item the list of alterations dominating A "strictly" (i.e. excluding those dominated by A)
    \item the strongly connected component to which A belongs 
    \item one "normal form", i.e. an alteration without strict dominators, which dominates A (directly or indirectly) 
  }
}
\details{
Given a network connecting samples and alterations, and given two groups of samples, red and blue, alteration A
will dominate alteration B if : 
\itemize{
 \item At least <redPercent>\% of the red samples affected by B are also affected by A, 
 \item At least <bluePercent>\% of the blue samples affected by A are also affected by B,
 \item The loci of A and B lie on the same chromosome, at most <distance> base pairs apart (this part is for alterations which are genomic, and can simply be ignored for other applications)
}
This is a somewhat more flexible version of the more natual criteria according to which A would dominate B if : 
\itemize{
 \item The red samples affected by B are also affected by A
 \item The blue samples affected by A are also affected by B
} 
To compute this simpler and more restrictive relation, use \code{redPercent=100} and \code{bluePercent=100}.
}
\examples{

#Real data
#bipartite graph data
data("tcga_bladder",package="musette")
graph.dele=matrix2graph(matrices$dele)
reds= (groups == 'basal')
names(reds)=names(groups)

# chromosome and position data
genes = unlist(rownames(matrices$dele))
chromosome=chromosome[genes]
position=position[genes]
dg = domination.graph(graph.dele,reds,chromosome=chromosome,locus=position)

dg$strict_dominated$CDKN2A
dg$strict_dominators$CDKN2A

}
