% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{encodeSolutionTree}
\alias{encodeSolutionTree}
\title{Create a Dataframe From the Solution Tree}
\usage{
encodeSolutionTree()
}
\value{
nothing.
}
\description{
This is an inner function, called by other R functions in this
package, but it is left available for the expert user or developer.
}
\details{
Gather information about the nodes contained in it and write them to
A dataframe. Note that it only contains numbers, which will later
have to be translated by R code into names, unknown to C++ code.
}
\seealso{
Other inner functions: 
\code{\link{connectedComponents}()},
\code{\link{exploreTree}()},
\code{\link{fullyGrowTree}()},
\code{\link{growTree}()},
\code{\link{rawDominationGraphFineTuned}()},
\code{\link{resetSolutionTree}()},
\code{\link{scc}()},
\code{\link{setBound}()},
\code{\link{setThreshold}()},
\code{\link{takeGraph}()},
\code{\link{takeGroups}()},
\code{\link{takeStepScore}()}
}
\concept{inner functions}
