% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{exploreTree}
\alias{exploreTree}
\title{Grow Solution Tree Until a Criterion is Met, forget the Solutions, keep the Growth history.}
\usage{
exploreTree()
}
\value{
A list showing the values successively taken by the step-score threshold, along
with the number of nodes and the number of leaves at each stage.
}
\description{
This is an inner function, called by other R functions in this
package, but it is left available for the expert user or developer.
}
\details{
The solution tree is grown using \code{\link{growTree}} until
its size reaches the set bound, or the step-score threshold reaches
the predefined value.
}
\seealso{
\code{\link{growTree}}, \code{\link{setThreshold}}, \code{\link{setBound}}

Other inner functions: 
\code{\link{connectedComponents}()},
\code{\link{encodeSolutionTree}()},
\code{\link{fullyGrowTree}()},
\code{\link{growTree}()},
\code{\link{rawDominationGraphFineTuned}()},
\code{\link{resetSolutionTree}()},
\code{\link{scc}()},
\code{\link{setBound}()},
\code{\link{setThreshold}()},
\code{\link{takeGraph}()},
\code{\link{takeGroups}()},
\code{\link{takeStepScore}()}
}
\concept{inner functions}
