% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fullyGrowTree}
\alias{fullyGrowTree}
\title{Grow Solution Tree Until a Criterion is Met}
\usage{
fullyGrowTree()
}
\value{
nothing.
}
\description{
This is an inner function, called by other R functions in this
package, but it is left available for the expert user or developer.
}
\details{
The solution tree is grown using \code{\link{growTree}} until
its size reaches the set bound, or the step-score threshold reaches
the predefined value.
}
\seealso{
\code{\link{growTree}}, \code{\link{setThreshold}}, \code{\link{setBound}}

Other inner functions: 
\code{\link{connectedComponents}()},
\code{\link{encodeSolutionTree}()},
\code{\link{exploreTree}()},
\code{\link{growTree}()},
\code{\link{rawDominationGraphFineTuned}()},
\code{\link{resetSolutionTree}()},
\code{\link{scc}()},
\code{\link{setBound}()},
\code{\link{setThreshold}()},
\code{\link{takeGraph}()},
\code{\link{takeGroups}()},
\code{\link{takeStepScore}()}
}
\concept{inner functions}
