% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutionTree.R
\name{initSolutionTree}
\alias{initSolutionTree}
\title{Initialize a Solution Tree}
\usage{
initSolutionTree(
  graph,
  reds,
  blues = !reds,
  threshold = 1,
  bound = 2000,
  stepmode = "strict"
)
}
\arguments{
\item{graph}{A named list of character vectors. Each of these vectors 
should have the name of one alteration, and contain the vector of the 
names of the samples affected by this alteration. Such a list can be 
obtained from a logical matrix using \code{\link{matrix2graph}}}

\item{reds}{A named logical vector whose names are the samples' names. The
values are \code{TRUE} for red samples, i.e. those which our solutions
should affect.}

\item{blues}{A similar vector specifying the blue samples, those that our
solutions should not affect. It is forbidden for a sample to be red and blue
at the same time !}

\item{threshold}{The maximum step-score for creating new solutions (see
details about the step-score in \link{musette-algorithm}).}

\item{bound}{The program will stop generating new solutions when this number
of solutions is reached. Note : if there are many \dQuote{equally good}
solutions they are added at the same time, possibly exceeding this
\code{bound}, then the algorithm stops.}

\item{stepmode}{A string to set how the step-score is calculated. Valid values
are \code{"original"}, \code{"unique"}, \code{"bestFirst"}, \code{"strict"}.}
}
\description{
Use the provided arguments to initialize data structures that
will contain, and help to build, a solution tree.
}
\details{
(See \link{musette-algorithm} to learn what "solution tree" means)
This is an internal function meant for use by other functions in the package. 
An expert user wanting to be in full control could use it, followed 
by calls to \link{growTree}, \link{fullyGrowTree}, or \link{exploreTree}.
}
