% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{matrix2graph}
\alias{matrix2graph}
\title{Make an Adjacency List out of a Logical Matrix}
\usage{
matrix2graph(matrix)
}
\arguments{
\item{matrix}{A logical matrix, with mandatory row and column names ! Row
names must correspond to alterations, and column names must correspond to
samples.}
}
\value{
An named list of character vectors. Its names are the alterations'
names, and each vector contains the names of the samples affected by the
corresponding alteration.
}
\description{
Produces an adjacency list of the type that can be passed to various
functions of package \link{musette}
}
\examples{
#Toy example
toy.matrix <- matrix(c(1,0,1,0,0,0,1,1,0,0,1,0,0,0,1,1,0,1,0,1,0,0,0,1),4,6,byrow=TRUE)
rownames(toy.matrix) <- LETTERS[1:4]
colnames(toy.matrix) <- letters[1:6]
toy.graph <- matrix2graph(toy.matrix)

#Real data
data("tcga_bladder",package="musette")
graph.dele=matrix2graph(matrices$dele)


}
