% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{musette2csv}
\alias{musette2csv}
\title{Turn  Vectors in a Dataframe into Strings for Better Rendering in a CSV File}
\usage{
musette2csv(d)
}
\arguments{
\item{d}{A dataframe}
}
\value{
The same dataframe, where cells that are vectors have been replaced by semicolon-punctuated strings.
}
\description{
Cells of the solution dataframe containing vectors are uneasy to read. This function solves the problem when exporting the dataframe to a csv file.
}
\examples{
dataframe=data.frame(person=c("sam","ben","alice"))
dataframe$food=list(c("cheese","egg","bacon"),c("tomato"),c("bread","ham"))
\dontrun{write.csv(dataframe,"file.csv")} # causes an error ("unimplemented type")
d2=musette2csv(dataframe)
\dontrun{write.csv(d2,"file.csv")} # no error

}
