% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{resetSolutionTree}
\alias{resetSolutionTree}
\title{Reset the Solution Tree}
\usage{
resetSolutionTree()
}
\value{
nothing.
}
\description{
This is an inner function, called by other R functions in this
package, but it is left available for the expert user or developer.
}
\details{
Create a fresh solution tree, empty and ready to be grown, 
clearing the previously built solution tree, if necessary.
}
\seealso{
Other inner functions: 
\code{\link{connectedComponents}()},
\code{\link{encodeSolutionTree}()},
\code{\link{exploreTree}()},
\code{\link{fullyGrowTree}()},
\code{\link{growTree}()},
\code{\link{rawDominationGraphFineTuned}()},
\code{\link{scc}()},
\code{\link{setBound}()},
\code{\link{setThreshold}()},
\code{\link{takeGraph}()},
\code{\link{takeGroups}()},
\code{\link{takeStepScore}()}
}
\concept{inner functions}
