% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setThreshold}
\alias{setThreshold}
\title{Pass a Step-score Threshold to the C++ Code}
\usage{
setThreshold(threshold)
}
\arguments{
\item{threshold}{: An integer}
}
\value{
nothing.
}
\description{
This is an inner function, called by other R functions in this
package, but it is left available for the expert user or developer.
}
\seealso{
Other inner functions: 
\code{\link{connectedComponents}()},
\code{\link{encodeSolutionTree}()},
\code{\link{exploreTree}()},
\code{\link{fullyGrowTree}()},
\code{\link{growTree}()},
\code{\link{rawDominationGraphFineTuned}()},
\code{\link{resetSolutionTree}()},
\code{\link{scc}()},
\code{\link{setBound}()},
\code{\link{takeGraph}()},
\code{\link{takeGroups}()},
\code{\link{takeStepScore}()}
}
\concept{inner functions}
