% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumors.R
\name{sharedPathways}
\alias{sharedPathways}
\title{Analysis of the solutions in terms of pathways}
\usage{
sharedPathways(sol, alterations)
}
\arguments{
\item{alterations}{an alteration dataframe returned by do.musette}

\item{solutions}{a solution dataframe returned by do.musette}
}
\value{
The returned data.frame contains a row for each couple solution/pathway such
that at least two alterations do concern genes belonging to the pathway.
Columns include :
\itemize{
\item sol: identification number of the solution in the \code{solutions} dataframe
\item alterations describing the solution 
\item pathway
\item gene: alterations of the solution (including those dominated by the descriptors) that belong to the pathway
\item leader: leaders of the alterations in the gene list
\item red neighbors, blue neighbors of the alterations in the gene list, and the cardinal of those sets
}
}
\description{
Given the solutions and alterations 
dataframes returned by do.musette, it returns a new dataframe with solutions
reorganized by pathways shared by several genes in a common solution
}
\examples{
#Real bladder data, restricted on the 20 first solution sets including only mutations
data("tcga_bladder",package="musette")   
reds= (groups == 'basal')
names(reds)=names(groups)

ll=do.musette(matrices=matrices, reds=reds, blind_domination_step = c("ampli","dele"), 
color_domination_step = c("ampli","dele"), blind_distance=5000000, 
color_distance=5000000, blind_percent=90, red_percent=80, 
blue_percent=80, chromosome = chromosome, longname=longname,
pathways=pathways,position=position, bound=20)

spdf = sharedPathways(ll$solutions,ll$alterations)
View(spdf)


}
