% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunction.R
\name{solution.oncomatrix}
\alias{solution.oncomatrix}
\title{Generating an Oncoplot of a Given Solution}
\usage{
solution.oncomatrix(data, index, reds, display = TRUE)
}
\arguments{
\item{data}{A dataframe of solutions generated by \link{do.musette}}

\item{index}{The index of a solution in a \code{\link{solutions}} dataframe}

\item{reds}{A named logical vector whose names are the samples' names. The
values are \code{TRUE} for red samples, i.e. those which the solution
should affect.}

\item{display}{A boolean value indicating if the corresponding oncoplot should be displayed}
}
\value{
A matrix which can be used with function oncoPrint in package ComplexHeatmap to generate an oncoplot.
}
\description{
To be used with package ComplexHeatMap in order to visualize solutions as oncoplots.
}
\examples{
#Real bladder data, restricted on the 20 first solution sets including only mutations
data("tcga_bladder",package="musette")   
reds= (groups == 'basal')
names(reds)=names(groups)

ll=do.musette(matrices=matrices, reds=reds, blind_domination_step = c("ampli","dele"), 
color_domination_step = c("ampli","dele"), blind_distance=5000000, 
color_distance=5000000, blind_percent=90, red_percent=80, 
blue_percent=80, chromosome = chromosome, longname=longname,
pathways=pathways,position=position, bound=20)

#oncomatrix for the best solution found so far in terms of score

index <- rownames(ll$solutions)[1]   
solution.oncomatrix(ll,index,reds)


}
