% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{takeGraph}
\alias{takeGraph}
\title{Get Ready to Work with This Graph}
\usage{
takeGraph(data)
}
\arguments{
\item{data}{A vector of integer vectors. \code{data[[i]]} should be an integer vector containing the ids of the samples hit by alteration \code{i}}
}
\value{
nothing.
}
\description{
This is an inner function, called by other R functions in this
package, but it is left available for the expert user or developer.
}
\details{
Set Some Internal Global Variables to Deal with a Given Graph : 
the degree distribution of the alterations, the lists of neighbours,
the number of alterations, and the memo that will contain score
distribution tables used for step-score computation.
}
\seealso{
Other inner functions: 
\code{\link{connectedComponents}()},
\code{\link{encodeSolutionTree}()},
\code{\link{exploreTree}()},
\code{\link{fullyGrowTree}()},
\code{\link{growTree}()},
\code{\link{rawDominationGraphFineTuned}()},
\code{\link{resetSolutionTree}()},
\code{\link{scc}()},
\code{\link{setBound}()},
\code{\link{setThreshold}()},
\code{\link{takeGroups}()},
\code{\link{takeStepScore}()}
}
\concept{inner functions}
