% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{takeGroups}
\alias{takeGroups}
\title{Get Ready to Work with These Groups}
\usage{
takeGroups(red, blue)
}
\arguments{
\item{red}{A logical vector. \code{red[[i]]} should be true iff sample \code{i} is red}

\item{blue}{A logical vector. \code{blue[[i]]} should be true iff sample \code{i} is blue}
}
\value{
nothing.
}
\description{
This is an inner function, called by other R functions in this
package, but it is left available for the expert user or developer.
}
\details{
Set Some Internal Global Variables to Deal with a Given Grouping of
the Samples : number of red/blue samples and the groups themselves.
}
\seealso{
Other inner functions: 
\code{\link{connectedComponents}()},
\code{\link{encodeSolutionTree}()},
\code{\link{exploreTree}()},
\code{\link{fullyGrowTree}()},
\code{\link{growTree}()},
\code{\link{rawDominationGraphFineTuned}()},
\code{\link{resetSolutionTree}()},
\code{\link{scc}()},
\code{\link{setBound}()},
\code{\link{setThreshold}()},
\code{\link{takeGraph}()},
\code{\link{takeStepScore}()}
}
\concept{inner functions}
